+++ lib/concierge/suppliers/avantio/commands/accommodations_fetcher.rb
@@ -4,7 +4,7 @@ module Avantio
     #
     # Avantio provides accommodations' information in zipped file available by URL.
     # This class fetches the file and parses the accommodations to the
-    # hash of +Avantio::Entities::Accommodation+ grouped by GACode
+    # hash of +Avantio::Entities::Accommodation+ grouped by LoginGA
     #
     # Usage
     #
@@ -44,7 +44,7 @@ module Avantio
 
       def build_accommodations(accommodations_raw)
         accommodations = accommodations_raw.xpath('/AccommodationList/AccommodationData')
-        Array(accommodations).map { |accommodation| mapper.build(accommodation) }.group_by(&:ga_code)
+        Array(accommodations).map { |accommodation| mapper.build(accommodation) }.group_by(&:login_ga)
       end
     end
   end